Feature/feedback withdrawal

+++ src/main/java/com/example/project/auth/controller/AuthController.java
@@ -46,16 +46,22 @@ public class AuthController {
 
     @PutMapping("")
     public ResponseEntity<?> updateAuthInfo(@RequestBody @Valid UpdateAuthRequest updateAuthRequest, HttpServletRequest request) {
+        Boolean auth = authService.deleteAuthInfo(request, updateAuthRequest);
+
         if (updateAuthRequest.getType().equals("회원정보 수정")) {
             authService.updateAuthInfo(updateAuthRequest, request);
             return ResponseEntity.ok("회원정보가 수정되었습니다.");
-
-        } else if (updateAuthRequest.getType().equals("회원 탈퇴")) {
-            authService.deleteAuthInfo(request, updateAuthRequest);
-            return ResponseEntity.ok("회원탈퇴가 되었습니다.");
         }
-        throw new TypeNotFound();
+        if (updateAuthRequest.getType().equals("회원 탈퇴")) {
+            if (auth != null) {
+                authService.deleteAuthInfo(request, updateAuthRequest);
+                return ResponseEntity.ok("회원탈퇴가 되었습니다.");
+            }
+            throw new TypeNotFound();
+        }
+        return null;
     }
+
 
     @GetMapping("")
     public AuthView readAuth(HttpServletRequest request) {
